@@ -14,15 +14,13 @@
         "rss": "https://github.com/phly/phly-configfactory/releases.atom"
     },
     "require": {
-        "php": "^7.3 || ~8.0.0 || ~8.1.0",
-        "psr/container": "^1.0"
+        "php": "~8.0.0 || ~8.1.0 || ~8.2.0",
+        "psr/container": "^1.0 || ^2.0"
     },
     "require-dev": {
         "laminas/laminas-coding-standard": "~2.3.0",
-        "laminas/laminas-dependency-plugin": "^2.1",
         "laminas/laminas-servicemanager": "^3.4",
-        "phpspec/prophecy-phpunit": "^2.0",
-        "phpunit/phpunit": "^9.5.5"
+        "phpunit/phpunit": "^9.6"
     },
     "suggest": {
         "laminas/laminas-servicemanager": "Install laminas-servicemanager to use the ConfigAbstractFactory"
@@ -38,11 +36,12 @@
         }
     },
     "config": {
-        "sort-packages": true
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-master": "1.0.x-dev"
+        "sort-packages": true,
+        "platform": {
+            "php": "8.0.99"
+        },
+        "allow-plugins": {
+            "dealerdirect/phpcodesniffer-composer-installer": true
         }
     },
     "scripts": {
